+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # unreleased
 ## Features
 * Display number of search results
+* Automatically restrict search by commander's color identity (if applicable)
 
 ## Fixes
 * Fix issue where search would stall out when no search terms were provided